Title: 큰 숫자만 계속 고르기, Time: 22ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240909/큰 숫자만 계속 고르기/keep-picking-the-big-number.cpp
@@ -10,7 +10,7 @@ int main()
 	int N, M;
 	cin >> N >> M;
 
-	priority_queue<long long> pq;
+	priority_queue<int> pq;
 
 	for (int i = 0; i < N; i++)
 	{
